workshop-botzip/workshop-bot/requirements.txt
python-telegram-bot==20.7
gigachat==0.1.29
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
